Add : lombok plugin
