+++ build.gradle
@@ -28,6 +28,7 @@ dependencies {
     compile('pl.allegro.tech.boot:handlebars-spring-boot-starter:0.3.0')
     compile('org.springframework.boot:spring-boot-devtools')
     compile('org.hibernate:hibernate-java8')
+    compileOnly('org.projectlombok:lombok')
     runtime('com.h2database:h2')
     runtime('net.rakugakibox.spring.boot:logback-access-spring-boot-starter:2.7.1')
     testCompile('org.springframework.boot:spring-boot-starter-test')
