+++ src/main/java/codesquad/domain/question/Question.java
@@ -1,0 +1,33 @@
+package codesquad.domain.question;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Getter
+@Setter
+@Entity
+public class Question {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String writer;
+
+    private String title;
+
+    private String contents;
+
+    private String createdDate;
+
+    public void update(Question modifiedQuestion) {
+        this.writer = modifiedQuestion.writer;
+        this.title = modifiedQuestion.title;
+        this.contents = modifiedQuestion.contents;
+    }
+}
