+++ src/main/java/codesquad/domain/question/QuestionRepository.java
@@ -1,4 +1,4 @@
-package codesquad.domain;
+package codesquad.domain.question;
 
 import org.springframework.data.repository.CrudRepository;
